tidied up file struture
