--- afresh/app.py
@@ -1,4 +1,0 @@
-from application import app
-
-if __name__ == "__main__":
-    app.run(debug=True, host='0.0.0.0')
