--- afresh/application/__init__.py
@@ -1,13 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from os import getenv
-
-app = Flask(__name__)
-
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///data.db'
-app.config['SECRET_KEY'] = "abcd"
-
-db = SQLAlchemy(app)
-
-
-from application import routes
