--- afresh/application/templates/layout.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Academic Paper Index Page</title>
-</head>
-<body>
-    <div class="header">
-        <h1> Nonsense Academic Database</h1>
-        <div id = "links">
-            <a href="{{ url_for('index') }}">Home</a>
-            <a href="{{ url_for('update_academic') }}">Update an Academic's Details</a>
-            <a href="{{ url_for('update_paper') }}">Update a Papers Details</a>
-            <a href="{{ url_for('add_academic') }}">Add an Academic</a>
-            <a href="{{ url_for('add_paper') }}">Add a Paper</a>
-            <a href="{{ url_for('delete_paper') }}">Delete a Paper</a>
-            <a href="{{ url_for('delete_academic') }}">Delete an Academic</a>
-            <a href="{{ url_for('about') }}">About</a>
-        </div>
-    </div>
-    <div class="container">        
-        {% block body %}
-        {% endblock body %}
-    </div>
-    <div id="footer">
-        <p>Michael Corcoran</p>
-    </div>
-</body>
-</html>
