+++ application/templates/layout.html
@@ -8,9 +8,9 @@
 </head>
 <body>
     <div class="header">
-        <h1> {{ title_of+page}} </h1>
+        <h1> Nonsense Academic Database</h1>
         <div id = "links">
-            <a href="{{ url_for('home') }}">Home</a>}
+            <a href="{{ url_for('index') }}">Home</a>
             <a href="{{ url_for('update_academic') }}">Update an Academic's Details</a>
             <a href="{{ url_for('update_paper') }}">Update a Papers Details</a>
             <a href="{{ url_for('add_academic') }}">Add an Academic</a>
