--- application/templates/update_paper.html
@@ -1,4 +1,0 @@
-{% extends 'layout.html'%}
-
-{% block body %}
-{% endblock body %}
