+++ requirements.txt
@@ -1,19 +1,8 @@
-atomicwrites==1.4.0
-attrs==21.4.0
-click==8.0.3
-colorama==0.4.4
-Flask==2.0.2
-greenlet==1.1.2
-iniconfig==1.1.1
-itsdangerous==2.0.1
-Jinja2==3.0.3
-MarkupSafe==2.0.1
-packaging==21.3
-pluggy==1.0.0
-py==1.11.0
-PyMySQL==1.0.2
-pyparsing==3.0.7
-pytest==7.0.0
-SQLAlchemy==1.4.31
-tomli==2.0.1
-Werkzeug==2.0.3
+Flask
+Flask-SQLAlchemy
+Flask-WTF
+itsdangerous
+Jinja2
+MarkupSafe
+SQLAlchemy
+WTForms
